requirements.txt
azure-cli-core>=2.50.0
azure-identity>=1.15.0
azure-mgmt-netapp>=7.0.0
azure-mgmt-resource>=23.0.0
azure-mgmt-subscription>=3.1.1
pytest>=8.3.5